Refactor
As the project grew, some refactoring was needed:
- split main_window.py into main_window.py, plotting.py, and parameter_manager.py to keep files under ~1100 lines
- reduced function length and cyclomatic complexity by introducing helper functions
